Move code to Base handler class
