+++ inc/WebhookHandler/Base.php
@@ -9,6 +9,7 @@
 
 namespace Required\Traduttore\WebhookHandler;
 
+use Required\Traduttore\Project;
 use Required\Traduttore\WebhookHandler;
 use WP_REST_Request;
 
@@ -36,5 +37,49 @@ abstract class Base implements WebhookHandler {
 	 */
 	public function __construct( WP_REST_Request $request ) {
 		$this->request = $request;
+	}
+
+	/**
+	 * Returns the webhook sync secret.
+	 *
+	 * @since 3.0.0
+	 *
+	 * @param Project|null $project The current project if found.
+	 *
+	 * @return string Secret if set, null otherwise.
+	 */
+	protected function get_secret( Project $project = null ): ?string {
+		$secret = $project ? $project->get_repository_webhook_secret() : null;
+
+		switch ( get_class( $this ) ) {
+			case Bitbucket::class:
+				if ( defined( 'TRADUTTORE_BITBUCKET_SYNC_SECRET' ) ) {
+					$secret = TRADUTTORE_BITBUCKET_SYNC_SECRET;
+				}
+				break;
+			case GitHub::class:
+				if ( defined( 'TRADUTTORE_GITHUB_SYNC_SECRET' ) ) {
+					$secret = TRADUTTORE_GITHUB_SYNC_SECRET;
+				}
+
+				break;
+			case GitLab::class:
+				if ( defined( 'TRADUTTORE_GITLAB_SYNC_SECRET' ) ) {
+					$secret = TRADUTTORE_GITLAB_SYNC_SECRET;
+				}
+
+				break;
+		}
+
+		/**
+		 * Filters the sync secret for an incoming webhook request.
+		 *
+		 * @since 3.0.0
+		 *
+		 * @param string         $secret  Webhook sync secret.
+		 * @param WebhookHandler $handler The current webhook handler instance.
+		 * @param Project|null   $project The current project if found.
+		 */
+		return apply_filters( 'traduttore.webhook_secret', $secret, $this, $project );
 	}
 }
